refactor(topic-list): rename topic list v2 implementation
