+++ src/renderer/src/pages/home/Tabs/components/Topics.helpers.ts
